+++ src/types/BaseConfig.ts
@@ -4,5 +4,6 @@ export type BaseConfig = {
     baseUrl: string;
     cachePath: string;
     prefix: string;
+    cleanupLocalAttachmentFiles: boolean,
     pages: Page[];
 };